renamed main class of rawTextParser #660
